Reorganized hw_sim into separate files
