+++ profiles/rocket_profile.py
@@ -2,7 +2,6 @@ from abc import ABC, abstractmethod
 from typing import Dict, List
 from connections.connection import Connection
 from main_window.packet_parser import PacketParser
-from connections.sim.hw_sim import HWSim
 from main_window.device_manager import DeviceType
 from .label import Label
