+++ tests/integration_tests/test_sim.py
@@ -2,7 +2,8 @@ import pytest
 from pytest import approx
 from .integration_utils import test_app, valid_paramitrization, all_devices, all_profiles, only_flare, flush_packets
 from connections.sim.sim_connection import FirmwareNotFound
-from connections.sim.hw_sim import HWSim, SensorType
+from connections.sim.hw.hw_sim import HWSim
+from connections.sim.hw.sensors.sensor import SensorType
 from main_window.competition.comp_app import CompApp
 from main_window.data_entry_id import DataEntryIds, DataEntryValues
 from main_window.device_manager import DeviceType
